Highlight all PRs without changes requested
On repositories with an unprotected master branch, PRs do not get the
status 'Review required' by default. However, for repos with protected
branches PRs will always have either 'changes requested' or 'review
required', so changing to check for a missing 'changes requested' status
should be equal.
The only extra check that needs to be added is to guard this from also
being applied on regular issues, since that was previously handled by
filtering for 'review required'.