Ignore approved PRs when finding review required
When updating the check for 'review required' to not look for the string
'review required' specifically, I erroneously only negated the 'changes
requested' check. Of course, there are three states, not two, a PR can
have, so Approved PRs that were still building on Travis were
highlighted as yellow now.

Surely we will never find a fourth forgotten state as well.