Remove hack for highlighting 5/6 passes on backend
We added a hack to highlight PRs where only 5/6 build jobs had passed in
order to still highlight PRs where the coverage check had caused the PR
to get a red build status. A while after this, we changed those failure
states to neutral, which means this hack should no longer be needed. And
now this hack only causes PRs to become highlighted when one of the test
or lint jobs fail, which means they should in fact not be highlighted.