Update safari extension to new web extension
The solution is the same as before. We have a small shell that the
chrome extension lives in, however with these updates to the safari
extension we end up with having a lot of the chrome apis fully supported
per safari 14. This means also that we must use safari 14, which should
not be a problem.