Handle deferred content in PR list
GitHub has started loading "review required" / "changes requested"
information asynchronously in the pull request list. This detects the
lazy loaded content, and disables highlighting of rows that still have
deferred content.

A new run of the colorizer is scheduled to run again in 100 ms as long
as deferred content exists.

This should ensure that no row is incorrectly highlighted while loading
and that when all loading is complete, all rows are highlighted
correctly.