Only add update alert once
Now that we can run the colorizer in multiple passes, we got an issue
where checking for updates ran more than once.

This extracts update checking out of the colorizing pass, to make it
easier to run it only once. Also, the UI code is updated to actually
handle running more than once as well, so that the text inside the
update alert is instead mutated.