Stop breaking layout of repo issue/pr views
This plugin includes some custom styling to override the look of PRs in
a list, mainly to make the repo name take up less place when browsing
PRs for more than one repo at once. This has forever been breaking the
layout of a lot of other views in GitHub. With this bit of extra logic
to check what page we're actually on, we can stop breaking the layout on
all those other pages!