Update CSS variable color names
GitHub seems to have rolled out an update where they changed some of the
naming (the base color-scale names are the same, but specific vars are
now prefixed with fg or bg or border).

green-pr is a feature in GitHub pull request colorizer that was
previously removed, so I deleted the stale CSS now instead of updating
it.